KE2: implement basic function call extraction
