+++ java/kotlin-extractor2/src/main/kotlin/KotlinFileExtractor.kt
@@ -2704,26 +2704,6 @@ OLD: KE1
             )
         }
 
-        private fun extractMethodAccessWithoutArgs(
-            returnType: IrType,
-            locId: Label<DbLocation>,
-            enclosingCallable: Label<out DbCallable>,
-            callsiteParent: Label<out DbExprparent>,
-            childIdx: Int,
-            enclosingStmt: Label<out DbStmt>,
-            methodLabel: Label<out DbCallable>?
-        ) =
-            tw.getFreshIdLabel<DbMethodaccess>().also { id ->
-                val type = useType(returnType)
-
-                tw.writeExprs_methodaccess(id, type.javaResult.id, callsiteParent, childIdx)
-                tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                extractExprContext(id, locId, enclosingCallable, enclosingStmt)
-
-                // The caller should have warned about this before, so we don't repeat the warning here.
-                if (methodLabel != null) tw.writeCallableBinding(id, methodLabel)
-            }
-
         private val defaultConstructorMarkerClass by lazy {
             referenceExternalClass("kotlin.jvm.internal.DefaultConstructorMarker")
         }
@@ -2947,44 +2927,6 @@ OLD: KE1
         private fun isFunctionInvoke(callTarget: IrFunction, drType: IrSimpleType) =
             (drType.isFunctionOrKFunction() || drType.isSuspendFunctionOrKFunction()) &&
                 callTarget.name.asString() == OperatorNameConventions.INVOKE.asString()
-
-        private fun getCalleeMethodId(
-            callTarget: IrFunction,
-            drType: IrType?,
-            allowInstantiatedGenericMethod: Boolean
-        ): Label<out DbCallable>? {
-            if (callTarget.isLocalFunction())
-                return getLocallyVisibleFunctionLabels(callTarget).function
-
-            if (
-                allowInstantiatedGenericMethod &&
-                    drType is IrSimpleType &&
-                    !isUnspecialised(drType, logger)
-            ) {
-                val calleeIsInvoke = isFunctionInvoke(callTarget, drType)
-
-                val extractionMethod =
-                    if (calleeIsInvoke) getFunctionInvokeMethod(drType.arguments) else callTarget
-
-                return extractionMethod?.let {
-                    val typeArgs =
-                        if (calleeIsInvoke && drType.arguments.size > BuiltInFunctionArity.BIG_ARITY) {
-                            // Big arity `invoke` methods have a special implementation on JVM, they are
-                            // transformed to a call to
-                            // `kotlin.jvm.functions.FunctionN<out R>::invoke(vararg args: Any?)`, so we
-                            // only need to pass the type
-                            // argument for the return type. Additionally, the arguments are extracted
-                            // inside an array literal below.
-                            listOf(drType.arguments.last())
-                        } else {
-                            getDeclaringTypeArguments(callTarget, drType)
-                        }
-                    useFunction<DbCallable>(extractionMethod, typeArgs)
-                }
-            } else {
-                return useFunction<DbCallable>(callTarget)
-            }
-        }
 
         private fun getCalleeRealOverrideTarget(f: IrFunction): IrFunction {
             val target = f.target.realOverrideTarget
@@ -3271,56 +3213,6 @@ OLD: KE1
                     this is IrField && this.isStatic ||
                     this is IrEnumEntry
 
-        private fun extractCallValueArguments(
-            callId: Label<out DbExprparent>,
-            call: IrFunctionAccessExpression,
-            enclosingStmt: Label<out DbStmt>,
-            enclosingCallable: Label<out DbCallable>,
-            idxOffset: Int
-        ) =
-            extractCallValueArguments(
-                callId,
-                (0 until call.valueArgumentsCount).map { call.getValueArgument(it) },
-                enclosingStmt,
-                enclosingCallable,
-                idxOffset
-            )
-
-        private fun extractCallValueArguments(
-            callId: Label<out DbExprparent>,
-            valueArguments: List<IrExpression?>,
-            enclosingStmt: Label<out DbStmt>,
-            enclosingCallable: Label<out DbCallable>,
-            idxOffset: Int,
-            extractVarargAsArray: Boolean = false
-        ) {
-            var i = 0
-            valueArguments.forEach { arg ->
-                if (arg != null) {
-                    if (arg is IrVararg && !extractVarargAsArray) {
-                        arg.elements.forEachIndexed { varargNo, vararg ->
-                            extractVarargElement(
-                                vararg,
-                                enclosingCallable,
-                                callId,
-                                i + idxOffset + varargNo,
-                                enclosingStmt
-                            )
-                        }
-                        i += arg.elements.size
-                    } else {
-                        extractExpressionExpr(
-                            arg,
-                            enclosingCallable,
-                            callId,
-                            (i++) + idxOffset,
-                            enclosingStmt
-                        )
-                    }
-                }
-            }
-        }
-
         private fun findFunction(cls: IrClass, name: String): IrFunction? =
             cls.declarations.findSubType<IrFunction> { it.name.asString() == name }
 
@@ -3534,912 +3426,6 @@ OLD: KE1
                 "Array" -> true
                 else -> false
             }
-
-        private fun extractCall(
-            c: IrCall,
-            callable: Label<out DbCallable>,
-            stmtExprParent: StmtExprParent
-        ) {
-            with("call", c) {
-                val owner = getBoundSymbolOwner(c.symbol, c) ?: return
-                val target = tryReplaceSyntheticFunction(owner)
-
-                // The vast majority of types of call want an expr context, so make one available
-                // lazily:
-                val exprParent by lazy { stmtExprParent.expr(c, callable) }
-
-                val parent by lazy { exprParent.parent }
-
-                val idx by lazy { exprParent.idx }
-
-                val enclosingStmt by lazy { exprParent.enclosingStmt }
-
-                fun extractMethodAccess(
-                    syntacticCallTarget: IrFunction,
-                    extractMethodTypeArguments: Boolean = true,
-                    extractClassTypeArguments: Boolean = false
-                ) {
-                    val typeArgs =
-                        if (extractMethodTypeArguments)
-                            (0 until c.typeArgumentsCount)
-                                .map { c.getTypeArgument(it) }
-                                .requireNoNullsOrNull()
-                        else listOf()
-
-                    if (typeArgs == null) {
-                        logger.warn("Missing type argument in extractMethodAccess")
-                        return
-                    }
-
-                    extractRawMethodAccess(
-                        syntacticCallTarget,
-                        c,
-                        c.type,
-                        callable,
-                        parent,
-                        idx,
-                        enclosingStmt,
-                        (0 until c.valueArgumentsCount).map { c.getValueArgument(it) },
-                        c.dispatchReceiver,
-                        c.extensionReceiver,
-                        typeArgs,
-                        extractClassTypeArguments,
-                        c.superQualifierSymbol
-                    )
-                }
-
-                fun extractSpecialEnumFunction(fnName: String) {
-                    if (c.typeArgumentsCount != 1) {
-                        logger.errorElement("Expected to find exactly one type argument", c)
-                        return
-                    }
-
-                    val enumType = (c.getTypeArgument(0) as? IrSimpleType)?.classifier?.owner
-                    if (enumType == null) {
-                        logger.errorElement("Couldn't find type of enum type", c)
-                        return
-                    }
-
-                    if (enumType is IrClass) {
-                        val func =
-                            enumType.declarations.findSubType<IrFunction> {
-                                it.name.asString() == fnName
-                            }
-                        if (func == null) {
-                            logger.errorElement("Couldn't find function $fnName on enum type", c)
-                            return
-                        }
-
-                        extractMethodAccess(func, false)
-                    } else if (enumType is IrTypeParameter && enumType.isReified) {
-                        // A call to `enumValues<T>()` is being extracted, where `T` is a reified type
-                        // parameter of an `inline` function.
-                        // We can't generate a valid expression here, because we would need to know the
-                        // type of T on the call site.
-                        // TODO: replace error expression with something that better shows this
-                        // expression is unrepresentable.
-                        val id = tw.getFreshIdLabel<DbErrorexpr>()
-                        val type = useType(c.type)
-
-                        tw.writeExprs_errorexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        extractExprContext(id, tw.getLocation(c), callable, enclosingStmt)
-                    } else {
-                        logger.errorElement("Unexpected enum type rep ${enumType.javaClass}", c)
-                    }
-                }
-
-                fun binopReceiver(
-                    id: Label<out DbExpr>,
-                    receiver: IrExpression?,
-                    receiverDescription: String
-                ) {
-                    extractExprContext(id, tw.getLocation(c), callable, enclosingStmt)
-
-                    if (receiver == null) {
-                        logger.errorElement("$receiverDescription not found", c)
-                    } else {
-                        extractExpressionExpr(receiver, callable, id, 0, enclosingStmt)
-                    }
-                    if (c.valueArgumentsCount < 1) {
-                        logger.errorElement("No RHS found", c)
-                    } else {
-                        if (c.valueArgumentsCount > 1) {
-                            logger.errorElement("Extra arguments found", c)
-                        }
-                        val arg = c.getValueArgument(0)
-                        if (arg == null) {
-                            logger.errorElement("RHS null", c)
-                        } else {
-                            extractExpressionExpr(arg, callable, id, 1, enclosingStmt)
-                        }
-                    }
-                }
-
-                fun unaryopReceiver(
-                    id: Label<out DbExpr>,
-                    receiver: IrExpression?,
-                    receiverDescription: String
-                ) {
-                    extractExprContext(id, tw.getLocation(c), callable, enclosingStmt)
-
-                    if (receiver == null) {
-                        logger.errorElement("$receiverDescription not found", c)
-                    } else {
-                        extractExpressionExpr(receiver, callable, id, 0, enclosingStmt)
-                    }
-                    if (c.valueArgumentsCount > 0) {
-                        logger.errorElement("Extra arguments found", c)
-                    }
-                }
-
-                /**
-                 * Populate the lhs of a binary op from this call's dispatch receiver, and the rhs from
-                 * its sole argument.
-                 */
-                fun binopDisp(id: Label<out DbExpr>) {
-                    binopReceiver(id, c.dispatchReceiver, "Dispatch receiver")
-                }
-
-                fun binopExt(id: Label<out DbExpr>) {
-                    binopReceiver(id, c.extensionReceiver, "Extension receiver")
-                }
-
-                fun unaryopDisp(id: Label<out DbExpr>) {
-                    unaryopReceiver(id, c.dispatchReceiver, "Dispatch receiver")
-                }
-
-                fun unaryopExt(id: Label<out DbExpr>) {
-                    unaryopReceiver(id, c.extensionReceiver, "Extension receiver")
-                }
-
-                val dr = c.dispatchReceiver
-                when {
-                    isNumericFunction(
-                        target,
-                        "plus",
-                        "minus",
-                        "times",
-                        "div",
-                        "rem",
-                        "and",
-                        "or",
-                        "xor",
-                        "shl",
-                        "shr",
-                        "ushr"
-                    ) -> {
-                        val type = useType(c.type)
-                        val id: Label<out DbExpr> =
-                            when (val targetName = target.name.asString()) {
-                                "minus" -> {
-                                    val id = tw.getFreshIdLabel<DbSubexpr>()
-                                    tw.writeExprs_subexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "times" -> {
-                                    val id = tw.getFreshIdLabel<DbMulexpr>()
-                                    tw.writeExprs_mulexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "div" -> {
-                                    val id = tw.getFreshIdLabel<DbDivexpr>()
-                                    tw.writeExprs_divexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "rem" -> {
-                                    val id = tw.getFreshIdLabel<DbRemexpr>()
-                                    tw.writeExprs_remexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "and" -> {
-                                    val id = tw.getFreshIdLabel<DbAndbitexpr>()
-                                    tw.writeExprs_andbitexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "or" -> {
-                                    val id = tw.getFreshIdLabel<DbOrbitexpr>()
-                                    tw.writeExprs_orbitexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "xor" -> {
-                                    val id = tw.getFreshIdLabel<DbXorbitexpr>()
-                                    tw.writeExprs_xorbitexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "shl" -> {
-                                    val id = tw.getFreshIdLabel<DbLshiftexpr>()
-                                    tw.writeExprs_lshiftexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "shr" -> {
-                                    val id = tw.getFreshIdLabel<DbRshiftexpr>()
-                                    tw.writeExprs_rshiftexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "ushr" -> {
-                                    val id = tw.getFreshIdLabel<DbUrshiftexpr>()
-                                    tw.writeExprs_urshiftexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                else -> {
-                                    logger.errorElement("Unhandled binary target name: $targetName", c)
-                                    return
-                                }
-                            }
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        if (
-                            isFunction(
-                                target,
-                                "kotlin",
-                                "Byte or Short",
-                                { it == "Byte" || it == "Short" },
-                                "and",
-                                "or",
-                                "xor"
-                            )
-                        )
-                            binopExt(id)
-                        else binopDisp(id)
-                    }
-                    // != gets desugared into not and ==. Here we resugar it.
-                    c.origin == IrStatementOrigin.EXCLEQ &&
-                        isFunction(target, "kotlin", "Boolean", "not") &&
-                        c.valueArgumentsCount == 0 &&
-                        dr != null &&
-                        dr is IrCall &&
-                        isBuiltinCallInternal(dr, "EQEQ") -> {
-                        val id = tw.getFreshIdLabel<DbValueneexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_valueneexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, dr, callable, enclosingStmt)
-                    }
-                    c.origin == IrStatementOrigin.EXCLEQEQ &&
-                        isFunction(target, "kotlin", "Boolean", "not") &&
-                        c.valueArgumentsCount == 0 &&
-                        dr != null &&
-                        dr is IrCall &&
-                        isBuiltinCallInternal(dr, "EQEQEQ") -> {
-                        val id = tw.getFreshIdLabel<DbNeexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_neexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, dr, callable, enclosingStmt)
-                    }
-                    c.origin == IrStatementOrigin.EXCLEQ &&
-                        isFunction(target, "kotlin", "Boolean", "not") &&
-                        c.valueArgumentsCount == 0 &&
-                        dr != null &&
-                        dr is IrCall &&
-                        isBuiltinCallInternal(dr, "ieee754equals") -> {
-                        val id = tw.getFreshIdLabel<DbNeexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_neexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, dr, callable, enclosingStmt)
-                    }
-                    isFunction(target, "kotlin", "Boolean", "not") -> {
-                        val id = tw.getFreshIdLabel<DbLognotexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_lognotexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        unaryopDisp(id)
-                    }
-                    isNumericFunction(target, "inv", "unaryMinus", "unaryPlus") -> {
-                        val type = useType(c.type)
-                        val id: Label<out DbExpr> =
-                            when (val targetName = target.name.asString()) {
-                                "inv" -> {
-                                    val id = tw.getFreshIdLabel<DbBitnotexpr>()
-                                    tw.writeExprs_bitnotexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "unaryMinus" -> {
-                                    val id = tw.getFreshIdLabel<DbMinusexpr>()
-                                    tw.writeExprs_minusexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                "unaryPlus" -> {
-                                    val id = tw.getFreshIdLabel<DbPlusexpr>()
-                                    tw.writeExprs_plusexpr(id, type.javaResult.id, parent, idx)
-                                    id
-                                }
-                                else -> {
-                                    logger.errorElement("Unhandled unary target name: $targetName", c)
-                                    return
-                                }
-                            }
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        if (
-                            isFunction(
-                                target,
-                                "kotlin",
-                                "Byte or Short",
-                                { it == "Byte" || it == "Short" },
-                                "inv"
-                            )
-                        )
-                            unaryopExt(id)
-                        else unaryopDisp(id)
-                    }
-                    // We need to handle all the builtin operators defines in BuiltInOperatorNames in
-                    //     compiler/ir/ir.tree/src/org/jetbrains/kotlin/ir/IrBuiltIns.kt
-                    // as they can't be extracted as external dependencies.
-                    isBuiltinCallInternal(c, "less") -> {
-                        if (c.origin != IrStatementOrigin.LT) {
-                            logger.warnElement("Unexpected origin for LT: ${c.origin}", c)
-                        }
-                        val id = tw.getFreshIdLabel<DbLtexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_ltexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, c, callable, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "lessOrEqual") -> {
-                        if (c.origin != IrStatementOrigin.LTEQ) {
-                            logger.warnElement("Unexpected origin for LTEQ: ${c.origin}", c)
-                        }
-                        val id = tw.getFreshIdLabel<DbLeexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_leexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, c, callable, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "greater") -> {
-                        if (c.origin != IrStatementOrigin.GT) {
-                            logger.warnElement("Unexpected origin for GT: ${c.origin}", c)
-                        }
-                        val id = tw.getFreshIdLabel<DbGtexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_gtexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, c, callable, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "greaterOrEqual") -> {
-                        if (c.origin != IrStatementOrigin.GTEQ) {
-                            logger.warnElement("Unexpected origin for GTEQ: ${c.origin}", c)
-                        }
-                        val id = tw.getFreshIdLabel<DbGeexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_geexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, c, callable, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "EQEQ") -> {
-                        if (c.origin != IrStatementOrigin.EQEQ) {
-                            logger.warnElement("Unexpected origin for EQEQ: ${c.origin}", c)
-                        }
-                        val id = tw.getFreshIdLabel<DbValueeqexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_valueeqexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, c, callable, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "EQEQEQ") -> {
-                        if (c.origin != IrStatementOrigin.EQEQEQ) {
-                            logger.warnElement("Unexpected origin for EQEQEQ: ${c.origin}", c)
-                        }
-                        val id = tw.getFreshIdLabel<DbEqexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_eqexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, c, callable, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "ieee754equals") -> {
-                        if (c.origin != IrStatementOrigin.EQEQ) {
-                            logger.warnElement("Unexpected origin for ieee754equals: ${c.origin}", c)
-                        }
-                        val id = tw.getFreshIdLabel<DbEqexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_eqexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binOp(id, c, callable, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "CHECK_NOT_NULL") -> {
-                        if (c.origin != IrStatementOrigin.EXCLEXCL) {
-                            logger.warnElement("Unexpected origin for CHECK_NOT_NULL: ${c.origin}", c)
-                        }
-
-                        val id = tw.getFreshIdLabel<DbNotnullexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_notnullexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        unaryOp(id, c, callable, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "THROW_CCE") -> {
-                        // TODO
-                        logger.errorElement("Unhandled builtin", c)
-                    }
-                    isBuiltinCallInternal(c, "THROW_ISE") -> {
-                        // TODO
-                        logger.errorElement("Unhandled builtin", c)
-                    }
-                    isBuiltinCallInternal(c, "noWhenBranchMatchedException") -> {
-                        kotlinNoWhenBranchMatchedConstructor?.let {
-                            val locId = tw.getLocation(c)
-                            val thrownType = useSimpleTypeClass(it.parentAsClass, listOf(), false)
-                            val stmtParent = stmtExprParent.stmt(c, callable)
-                            val throwId = tw.getFreshIdLabel<DbThrowstmt>()
-                            tw.writeStmts_throwstmt(
-                                throwId,
-                                stmtParent.parent,
-                                stmtParent.idx,
-                                callable
-                            )
-                            tw.writeHasLocation(throwId, locId)
-                            val newExprId =
-                                extractNewExpr(
-                                    it,
-                                    null,
-                                    thrownType,
-                                    locId,
-                                    throwId,
-                                    0,
-                                    callable,
-                                    throwId
-                                )
-                            if (newExprId == null) {
-                                logger.errorElement(
-                                    "No ID for newExpr in noWhenBranchMatchedException",
-                                    c
-                                )
-                            } else {
-                                extractTypeAccess(thrownType, locId, newExprId, -3, callable, throwId)
-                            }
-                        }
-                    }
-                    isBuiltinCallInternal(c, "illegalArgumentException") -> {
-                        // TODO
-                        logger.errorElement("Unhandled builtin", c)
-                    }
-                    isBuiltinCallInternal(c, "ANDAND") -> {
-                        // TODO
-                        logger.errorElement("Unhandled builtin", c)
-                    }
-                    isBuiltinCallInternal(c, "OROR") -> {
-                        // TODO
-                        logger.errorElement("Unhandled builtin", c)
-                    }
-                    isFunction(target, "kotlin", "Any", "toString", true) -> {
-                        stringValueOfObjectMethod?.let {
-                            extractRawMethodAccess(
-                                it,
-                                c,
-                                c.type,
-                                callable,
-                                parent,
-                                idx,
-                                enclosingStmt,
-                                listOf(c.extensionReceiver),
-                                null,
-                                null
-                            )
-                        }
-                    }
-                    isBuiltinCallKotlin(c, "enumValues") -> {
-                        extractSpecialEnumFunction("values")
-                    }
-                    isBuiltinCallKotlin(c, "enumValueOf") -> {
-                        extractSpecialEnumFunction("valueOf")
-                    }
-                    isBuiltinCallKotlin(c, "arrayOfNulls") -> {
-                        val id = tw.getFreshIdLabel<DbArraycreationexpr>()
-                        val type = useType(c.type)
-                        tw.writeExprs_arraycreationexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        val locId = tw.getLocation(c)
-                        extractExprContext(id, locId, callable, enclosingStmt)
-
-                        if (c.typeArgumentsCount == 1) {
-                            val typeArgument = c.getTypeArgument(0)
-                            if (typeArgument == null) {
-                                logger.errorElement("Type argument missing in an arrayOfNulls call", c)
-                            } else {
-                                extractTypeAccessRecursive(
-                                    typeArgument,
-                                    locId,
-                                    id,
-                                    -1,
-                                    callable,
-                                    enclosingStmt,
-                                    TypeContext.GENERIC_ARGUMENT
-                                )
-                            }
-                        } else {
-                            logger.errorElement(
-                                "Expected to find exactly one type argument in an arrayOfNulls call",
-                                c
-                            )
-                        }
-
-                        if (c.valueArgumentsCount == 1) {
-                            val dim = c.getValueArgument(0)
-                            if (dim != null) {
-                                extractExpressionExpr(dim, callable, id, 0, enclosingStmt)
-                            } else {
-                                logger.errorElement(
-                                    "Expected to find non-null argument in an arrayOfNulls call",
-                                    c
-                                )
-                            }
-                        } else {
-                            logger.errorElement(
-                                "Expected to find only one argument in an arrayOfNulls call",
-                                c
-                            )
-                        }
-                    }
-                    isBuiltinCallKotlin(c, "arrayOf") ||
-                        isBuiltinCallKotlin(c, "doubleArrayOf") ||
-                        isBuiltinCallKotlin(c, "floatArrayOf") ||
-                        isBuiltinCallKotlin(c, "longArrayOf") ||
-                        isBuiltinCallKotlin(c, "intArrayOf") ||
-                        isBuiltinCallKotlin(c, "charArrayOf") ||
-                        isBuiltinCallKotlin(c, "shortArrayOf") ||
-                        isBuiltinCallKotlin(c, "byteArrayOf") ||
-                        isBuiltinCallKotlin(c, "booleanArrayOf") -> {
-
-                        val isPrimitiveArrayCreation = !isBuiltinCallKotlin(c, "arrayOf")
-                        val elementType =
-                            if (isPrimitiveArrayCreation) {
-                                c.type.getArrayElementType(pluginContext.irBuiltIns)
-                            } else {
-                                // TODO: is there any reason not to always use getArrayElementType?
-                                if (c.typeArgumentsCount == 1) {
-                                    c.getTypeArgument(0).also {
-                                        if (it == null) {
-                                            logger.errorElement(
-                                                "Type argument missing in an arrayOf call",
-                                                c
-                                            )
-                                        }
-                                    }
-                                } else {
-                                    logger.errorElement(
-                                        "Expected to find one type argument in arrayOf call",
-                                        c
-                                    )
-                                    null
-                                }
-                            }
-
-                        val arg =
-                            if (c.valueArgumentsCount == 1) c.getValueArgument(0)
-                                else {
-                                    logger.errorElement(
-                                        "Expected to find only one (vararg) argument in ${c.symbol.owner.name.asString()} call",
-                                        c
-                                    )
-                                    null
-                                }
-                                ?.let {
-                                    if (it is IrVararg) it
-                                    else {
-                                        logger.errorElement(
-                                            "Expected to find vararg argument in ${c.symbol.owner.name.asString()} call",
-                                            c
-                                        )
-                                        null
-                                    }
-                                }
-
-                        extractArrayCreation(
-                            arg,
-                            c.type,
-                            elementType,
-                            isPrimitiveArrayCreation,
-                            c,
-                            parent,
-                            idx,
-                            callable,
-                            enclosingStmt
-                        )
-                    }
-                    isBuiltinCall(c, "<get-java>", "kotlin.jvm") -> {
-                        // Special case for KClass<*>.java, which is used in the Parcelize plugin. In
-                        // normal cases, this is already rewritten to the property referenced below:
-                        findTopLevelPropertyOrWarn(
-                                "kotlin.jvm",
-                                "java",
-                                "kotlin.jvm.JvmClassMappingKt",
-                                c
-                            )
-                            ?.let { javaProp ->
-                                val getter = javaProp.getter
-                                if (getter == null) {
-                                    logger.error(
-                                        "Couldn't find getter of `kotlin.jvm.JvmClassMappingKt::java`"
-                                    )
-                                    return
-                                }
-
-                                val ext = c.extensionReceiver
-                                if (ext == null) {
-                                    logger.errorElement(
-                                        "No extension receiver found for `KClass::java` call",
-                                        c
-                                    )
-                                    return
-                                }
-
-                                val argType =
-                                    (ext.type as? IrSimpleType)?.arguments?.firstOrNull()?.typeOrNull
-                                val typeArguments = if (argType == null) listOf() else listOf(argType)
-
-                                extractRawMethodAccess(
-                                    getter,
-                                    c,
-                                    c.type,
-                                    callable,
-                                    parent,
-                                    idx,
-                                    enclosingStmt,
-                                    listOf(),
-                                    null,
-                                    ext,
-                                    typeArguments
-                                )
-                            }
-                    }
-                    isFunction(
-                        target,
-                        "kotlin",
-                        "(some array type)",
-                        { isArrayType(it) },
-                        "iterator"
-                    ) -> {
-                        val parentClass = target.parent
-                        if (parentClass !is IrClass) {
-                            logger.errorElement("Iterator parent is not a class", c)
-                            return
-                        }
-
-                        var typeFilter =
-                            if (isGenericArrayType(parentClass.name.asString())) {
-                                "kotlin.jvm.internal.ArrayIteratorKt"
-                            } else {
-                                "kotlin.jvm.internal.ArrayIteratorsKt"
-                            }
-
-                        findTopLevelFunctionOrWarn(
-                                "kotlin.jvm.internal",
-                                "iterator",
-                                typeFilter,
-                                arrayOf(parentClass.kotlinFqName.asString()),
-                                c
-                            )
-                            ?.let { iteratorFn ->
-                                val dispatchReceiver = c.dispatchReceiver
-                                if (dispatchReceiver == null) {
-                                    logger.errorElement(
-                                        "No dispatch receiver found for array iterator call",
-                                        c
-                                    )
-                                } else {
-                                    val drType = dispatchReceiver.type
-                                    if (drType !is IrSimpleType) {
-                                        logger.errorElement(
-                                            "Dispatch receiver with unexpected type rep found for array iterator call: ${drType.javaClass}",
-                                            c
-                                        )
-                                    } else {
-                                        val typeArgs =
-                                            drType.arguments.map {
-                                                when (it) {
-                                                    is IrTypeProjection -> it.type
-                                                    else -> pluginContext.irBuiltIns.anyNType
-                                                }
-                                            }
-                                        extractRawMethodAccess(
-                                            iteratorFn,
-                                            c,
-                                            c.type,
-                                            callable,
-                                            parent,
-                                            idx,
-                                            enclosingStmt,
-                                            listOf(c.dispatchReceiver),
-                                            null,
-                                            null,
-                                            typeArgs
-                                        )
-                                    }
-                                }
-                            }
-                    }
-                    isFunction(target, "kotlin", "(some array type)", { isArrayType(it) }, "get") &&
-                        c.origin == IrStatementOrigin.GET_ARRAY_ELEMENT &&
-                        c.dispatchReceiver != null -> {
-                        val id = tw.getFreshIdLabel<DbArrayaccess>()
-                        val type = useType(c.type)
-                        tw.writeExprs_arrayaccess(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        binopDisp(id)
-                    }
-                    isFunction(target, "kotlin", "(some array type)", { isArrayType(it) }, "set") &&
-                        c.origin == IrStatementOrigin.EQ &&
-                        c.dispatchReceiver != null -> {
-                        val array = c.dispatchReceiver
-                        val arrayIdx = c.getValueArgument(0)
-                        val assignedValue = c.getValueArgument(1)
-
-                        if (array != null && arrayIdx != null && assignedValue != null) {
-
-                            val locId = tw.getLocation(c)
-                            extractAssignExpr(c.type, locId, parent, idx, callable, enclosingStmt)
-                                .also { assignId ->
-                                    tw.getFreshIdLabel<DbArrayaccess>().also { arrayAccessId ->
-                                        val arrayType = useType(array.type)
-                                        tw.writeExprs_arrayaccess(
-                                            arrayAccessId,
-                                            arrayType.javaResult.id,
-                                            assignId,
-                                            0
-                                        )
-                                        tw.writeExprsKotlinType(
-                                            arrayAccessId,
-                                            arrayType.kotlinResult.id
-                                        )
-                                        extractExprContext(
-                                            arrayAccessId,
-                                            locId,
-                                            callable,
-                                            enclosingStmt
-                                        )
-
-                                        extractExpressionExpr(
-                                            array,
-                                            callable,
-                                            arrayAccessId,
-                                            0,
-                                            enclosingStmt
-                                        )
-                                        extractExpressionExpr(
-                                            arrayIdx,
-                                            callable,
-                                            arrayAccessId,
-                                            1,
-                                            enclosingStmt
-                                        )
-                                    }
-                                    extractExpressionExpr(
-                                        assignedValue,
-                                        callable,
-                                        assignId,
-                                        1,
-                                        enclosingStmt
-                                    )
-                                }
-                        } else {
-                            logger.errorElement("Unexpected Array.set function signature", c)
-                        }
-                    }
-                    isBuiltinCall(c, "<unsafe-coerce>", "kotlin.jvm.internal") -> {
-
-                        if (c.valueArgumentsCount != 1) {
-                            logger.errorElement(
-                                "Expected to find one argument for a kotlin.jvm.internal.<unsafe-coerce>() call, but found ${c.valueArgumentsCount}",
-                                c
-                            )
-                            return
-                        }
-
-                        if (c.typeArgumentsCount != 2) {
-                            logger.errorElement(
-                                "Expected to find two type arguments for a kotlin.jvm.internal.<unsafe-coerce>() call, but found ${c.typeArgumentsCount}",
-                                c
-                            )
-                            return
-                        }
-                        val valueArg = c.getValueArgument(0)
-                        if (valueArg == null) {
-                            logger.errorElement(
-                                "Cannot find value argument for a kotlin.jvm.internal.<unsafe-coerce>() call",
-                                c
-                            )
-                            return
-                        }
-                        val typeArg = c.getTypeArgument(1)
-                        if (typeArg == null) {
-                            logger.errorElement(
-                                "Cannot find type argument for a kotlin.jvm.internal.<unsafe-coerce>() call",
-                                c
-                            )
-                            return
-                        }
-
-                        val id = tw.getFreshIdLabel<DbUnsafecoerceexpr>()
-                        val locId = tw.getLocation(c)
-                        val type = useType(c.type)
-                        tw.writeExprs_unsafecoerceexpr(id, type.javaResult.id, parent, idx)
-                        tw.writeExprsKotlinType(id, type.kotlinResult.id)
-                        extractExprContext(id, locId, callable, enclosingStmt)
-                        extractTypeAccessRecursive(typeArg, locId, id, 0, callable, enclosingStmt)
-                        extractExpressionExpr(valueArg, callable, id, 1, enclosingStmt)
-                    }
-                    isBuiltinCallInternal(c, "dataClassArrayMemberToString") -> {
-                        val arrayArg = c.getValueArgument(0)
-                        val realArrayClass = arrayArg?.type?.classOrNull
-                        if (realArrayClass == null) {
-                            logger.errorElement(
-                                "Argument to dataClassArrayMemberToString not a class",
-                                c
-                            )
-                            return
-                        }
-                        val realCallee =
-                            javaUtilArrays?.declarations?.findSubType<IrFunction> { decl ->
-                                decl.name.asString() == "toString" &&
-                                    decl.valueParameters.size == 1 &&
-                                    decl.valueParameters[0].type.classOrNull?.let {
-                                        it == realArrayClass
-                                    } == true
-                            }
-                        if (realCallee == null) {
-                            logger.errorElement(
-                                "Couldn't find a java.lang.Arrays.toString method matching class ${realArrayClass.owner.name}",
-                                c
-                            )
-                        } else {
-                            extractRawMethodAccess(
-                                realCallee,
-                                c,
-                                c.type,
-                                callable,
-                                parent,
-                                idx,
-                                enclosingStmt,
-                                listOf(arrayArg),
-                                null,
-                                null
-                            )
-                        }
-                    }
-                    isBuiltinCallInternal(c, "dataClassArrayMemberHashCode") -> {
-                        val arrayArg = c.getValueArgument(0)
-                        val realArrayClass = arrayArg?.type?.classOrNull
-                        if (realArrayClass == null) {
-                            logger.errorElement(
-                                "Argument to dataClassArrayMemberHashCode not a class",
-                                c
-                            )
-                            return
-                        }
-                        val realCallee =
-                            javaUtilArrays?.declarations?.findSubType<IrFunction> { decl ->
-                                decl.name.asString() == "hashCode" &&
-                                    decl.valueParameters.size == 1 &&
-                                    decl.valueParameters[0].type.classOrNull?.let {
-                                        it == realArrayClass
-                                    } == true
-                            }
-                        if (realCallee == null) {
-                            logger.errorElement(
-                                "Couldn't find a java.lang.Arrays.hashCode method matching class ${realArrayClass.owner.name}",
-                                c
-                            )
-                        } else {
-                            extractRawMethodAccess(
-                                realCallee,
-                                c,
-                                c.type,
-                                callable,
-                                parent,
-                                idx,
-                                enclosingStmt,
-                                listOf(arrayArg),
-                                null,
-                                null
-                            )
-                        }
-                    }
-                    else -> {
-                        extractMethodAccess(target, true, true)
-                    }
-                }
-            }
-        }
 
         private fun extractArrayCreation(
             elementList: IrVararg?,
